Add Events docs
This commit introduces an events docs, which lists the interesting
things that are happening in the Berlin SRE community.

== Design Notes
=== Events.adoc

The doc was originally named "Conferences", however the author (at least
had) intended on starting a Meetup. Meetup != conference, and thus
"Events" was born.

=== Naming structure

The thinking is that users will be "reading ahead" the current events,
and will look for those they can attend by dates than topics.